+++ src/main/kotlin/com/oaalto/agent/settings/AgentSettingsDetailPanel.kt
@@ -4,11 +4,17 @@ import com.intellij.ui.components.JBLabel
 import com.intellij.ui.table.JBTable
 import javax.swing.JCheckBox
 import javax.swing.JComponent
+import javax.swing.JPanel
 
 internal data class DetailPanelHints(
     val mcpScopeHint: JBLabel,
     val mcpPluginsHint: JBLabel,
     val envHint: JBLabel,
+)
+
+internal data class ObservationHelpPanel(
+    val panel: JPanel,
+    val hasBasePath: Boolean,
 )
 
 internal data class EnvironmentVariablesEditor(